requirements.txt
httpx
aiofiles
PyYAML
tqdm
git+https://github.com/dongresource/UnityPackFF.git@ec9c4524ffc93e0482c500d810dd63cfa110be40
